+++ packages/identity-x/browser/errors/authentication.js
