+++ packages/identity-x/browser/errors/feature.js
@@ -1,5 +1,5 @@
 export default class FeatureError extends Error {
-  constructor(message, code = 500) {
+  constructor(message, code = 400) {
     super(message);
     this.name = 'FeatureError';
     this.code = code;
